# internal/transport/http/docs/swagger.yaml
definitions:
  dto.CreateUserRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role:
        type: string
    type: object
  dto.UserResponse:
    properties:
      createdAt:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        type: string
      jwt:
        type: string
      phone_number:
        type: string
      role:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/users:
    post:
      consumes:
      - application/json
      description: Create a new user. Role will be blank/ignored for public creation.
      parameters:
      - description: Create User Request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a user (public)
      tags:
      - Users
  /v1/users/{userId}:
    delete:
      description: Delete a user by ID. Requires authentication.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a user
      tags:
      - Users
  /v1/users/admin:
    post:
      consumes:
      - application/json
      description: Create a new user as an administrator. Requires authentication.
      parameters:
      - description: Create User Request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a user (admin)
      tags:
      - Users
swagger: "2.0"